+++ src/NHA/NHA.php
@@ -66,7 +66,7 @@ use NHA\Repository\WorldRepository;
  * @property SocialRepository         $social
  * @property WorldRepository          $world
  *
- * @version 3.1.16
+ * @version 3.1.17
  */
 class NHA extends MessageCommandClient
 {